requirements.txt
streamlit
groq
langchain
langchain-groq
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
edge-tts
requests